Export Geolocation function to a Provider
