+++ src/app/app.component.ts
@@ -1,9 +1,9 @@
+import { GpsproviderProvider } from './../providers/gpsprovider/gpsprovider';
 import { Component, ViewChild } from '@angular/core';
 import { Platform, Nav, Config } from 'ionic-angular';
 
 import { StatusBar } from '@ionic-native/status-bar';
 import { SplashScreen } from '@ionic-native/splash-screen';
-import { Geolocation } from '@ionic-native/geolocation';
 
 import { CardsPage } from '../pages/cards/cards';
 import { ContentPage } from '../pages/content/content';
@@ -62,9 +62,8 @@ export class MyApp {
     { title: 'Search', component: SearchPage }
   ]
 
-  constructor(private translate: TranslateService, private platform: Platform, settings: Settings, private config: Config, private statusBar: StatusBar, private splashScreen: SplashScreen, private geolocation: Geolocation) {
+  constructor(private translate: TranslateService, private platform: Platform, settings: Settings, private config: Config, private statusBar: StatusBar, private splashScreen: SplashScreen) {
     this.initTranslate();
-    //this.getGeolocation();
   }
 
   ionViewDidLoad() {
@@ -95,33 +94,5 @@ export class MyApp {
     // Reset the content nav to have just this page
     // we wouldn't want the back button to show in this scenario
     this.nav.setRoot(page.component);
-  }
-
-  getGeolocation() {
-    this.geolocation.getCurrentPosition().then((resp) => {
-      console.log(resp.coords.latitude);
-      console.log(resp.coords.longitude);
-     }).catch((error) => {
-       console.log('Error getting location', error);
-     });
-  }
-
-  calculateETA(coords_a, coords_b) {
-    var lat_delta = coords_a.latitude - coords_b.latitude;
-    var lng_delta = coords_a.longitude - coords_b.longitude;
-    return lat_delta * lat_delta + lng_delta * lng_delta;
-  }
-
-  watchGeolocation(coords, max_time, func) {
-     let watch = this.geolocation.watchPosition();
-     watch.subscribe((data) => {
-      // data can be a set of coordinates, or an error (if an error occurred).
-      // data.coords.latitude
-      // data.coords.longitude
-      if (!(data instanceof Error))
-        if (this.calculateETA(data.coords, coords) < max_time) {
-          func();
-        }
-     });
   }
 }
