+++ src/app/app.module.ts
@@ -34,6 +34,7 @@ import { Geolocation } from '@ionic-native/geolocation';
 
 import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import { GpsproviderProvider } from '../providers/gpsprovider/gpsprovider';
 
 // The translate loader needs to know where to load i18n files
 // in Ionic's static asset pipeline.
@@ -116,7 +117,8 @@ export function provideSettings(storage: Storage) {
     Geolocation,
     { provide: Settings, useFactory: provideSettings, deps: [Storage] },
     // Keep this to enable Ionic's runtime error handling during development
-    { provide: ErrorHandler, useClass: IonicErrorHandler }
+    { provide: ErrorHandler, useClass: IonicErrorHandler },
+    GpsproviderProvider
   ]
 })
 export class AppModule { }
